pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nyc-tlc-analytics-warehouse"
version = "0.1.0"
description = "Data warehouse project for e-commerce behavior analytics"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Data Engineering Team" }]
dependencies = [
  "google-cloud-storage>=2.16.0",
  "google-cloud-bigquery>=3.20.0",
  "google-cloud-bigquery-storage>=2.25.0",
  "pyarrow>=16.0.0",
  "python-dotenv>=1.0.1",
  "pyspark>=3.4.0",
  "kaggle>=1.6.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
